requirements.txt
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
faiss-cpu
duckduckgo-search
google-generativeai
fastapi[standard]
sqlalchemy
google-genai


unstructured
markdown

## Search Modules
googlesearch-python
pycountry
wikipedia

## Vector DBs
# chromadb 
# lancedb some problem in my pc
aiofiles
langchain_postgres
pgvector
psycopg2
langchain-astradb

## App
python-jose[cryptography]

## Streamlit
streamlit>=1.42.0
Authlib>=1.3.2